% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaViz.R
\docType{methods}
\name{createUI,TrenaViz-method}
\alias{createUI,TrenaViz-method}
\alias{createUI}
\title{Create the user interface}
\usage{
\S4method{createUI}{TrenaViz}(obj)
}
\arguments{
\item{obj}{An object of class TrenaViz}
}
\description{
Create the user interface
}
